Remove cpu.c and move functions to cpu.h
